fixup! make transport testing reusable
