+++ pkg/storage/closedts/transport/transport_util_test.go
@@ -16,12 +16,12 @@ package transport_test
 
 import (
 	"context"
-	"io"
 
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/settings/cluster"
 	"github.com/cockroachdb/cockroach/pkg/storage/closedts/ctpb"
 	"github.com/cockroachdb/cockroach/pkg/storage/closedts/transport"
+	"github.com/cockroachdb/cockroach/pkg/storage/closedts/transport/testutils"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 )
@@ -33,7 +33,7 @@ type TestContainer struct {
 	Notifyee  *TestNotifyee
 	Refreshed *RefreshTracker
 	Server    *transport.Server
-	Dialer    *Dialer
+	Dialer    *testutils.ChanDialer
 	Clients   *transport.Clients
 }
 
@@ -88,113 +88,6 @@ func (tn *TestNotifyee) Notify(nodeID roachpb.NodeID) chan<- ctpb.Entry {
 		}
 	})
 	return ch
-}
-
-type client struct {
-	ctx     context.Context
-	stopper *stop.Stopper
-	send    chan *ctpb.Reaction
-	recv    chan *ctpb.Entry
-
-	observe func(interface{})
-}
-
-func (c *client) Send(msg *ctpb.Reaction) error {
-	select {
-	case <-c.stopper.ShouldQuiesce():
-		return io.EOF
-	case c.send <- msg:
-		c.observe(msg)
-		return nil
-	}
-}
-
-func (c *client) Recv() (*ctpb.Entry, error) {
-	select {
-	case <-c.stopper.ShouldQuiesce():
-		return nil, io.EOF
-	case msg := <-c.recv:
-		c.observe(msg)
-		return msg, nil
-	}
-}
-
-func (c *client) CloseSend() error {
-	close(c.send)
-	return nil
-}
-
-func (c *client) Context() context.Context {
-	return c.ctx
-}
-
-type incomingClient client
-
-func (c *incomingClient) Send(msg *ctpb.Entry) error {
-	select {
-	case <-c.stopper.ShouldQuiesce():
-		return io.EOF
-	case c.recv <- msg:
-		return nil
-	}
-}
-
-func (c *incomingClient) Recv() (*ctpb.Reaction, error) {
-	select {
-	case <-c.stopper.ShouldQuiesce():
-		return nil, io.EOF
-	case msg, ok := <-c.send:
-		if !ok {
-			return nil, io.EOF
-		}
-		return msg, nil
-	}
-}
-
-func (c *incomingClient) Context() context.Context {
-	return c.ctx
-}
-
-type Dialer struct {
-	stopper *stop.Stopper
-	server  *transport.Server
-
-	mu struct {
-		syncutil.Mutex
-		transcripts map[roachpb.NodeID][]interface{}
-	}
-}
-
-func (d *Dialer) transcript(nodeID roachpb.NodeID) []interface{} {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	return append([]interface{}(nil), d.mu.transcripts[nodeID]...)
-}
-
-func (d *Dialer) Dial(ctx context.Context, nodeID roachpb.NodeID) (ctpb.Client, error) {
-	c := &client{
-		ctx:     ctx,
-		send:    make(chan *ctpb.Reaction),
-		recv:    make(chan *ctpb.Entry),
-		stopper: d.stopper,
-		observe: func(msg interface{}) {
-			d.mu.Lock()
-			if d.mu.transcripts == nil {
-				d.mu.transcripts = map[roachpb.NodeID][]interface{}{}
-			}
-			d.mu.transcripts[nodeID] = append(d.mu.transcripts[nodeID], msg)
-			d.mu.Unlock()
-		},
-	}
-
-	d.stopper.RunWorker(ctx, func(ctx context.Context) {
-		_ = d.server.Handle((*incomingClient)(c))
-	})
-	return c, nil
-
-}
-func (d *Dialer) Ready(nodeID roachpb.NodeID) bool {
-	return true
 }
 
 type RefreshTracker struct {
